+++ src/components/general/CookieBanner/cookieBanner.js
@@ -1,22 +1,19 @@
 import './cookieBanner.html.twig';
 import './cookieBanner.scss';
+import {domContentLoadedWrapper} from '../../../domContentLoadedWrapper';
 
 export default function () {
-  let DOMContentFirstLoad = true;
-
-  window.addEventListener('DOMContentLoaded', function () {
-    if (DOMContentFirstLoad) {
-      DOMContentFirstLoad = false;
-      return;
-    }
-
+  function callback() {
     const acceptButton = document.querySelector('.cookie-banner__accept');
     if (!acceptButton) return;
 
     acceptButton.addEventListener('click', () => {
       const cookieBanner = document.querySelector('.cookie-banner');
-      if (!cookieBanner) return;
-      cookieBanner.classList.add('cookie-banner--hidden');
+      if (cookieBanner) {
+        cookieBanner.classList.add('cookie-banner--hidden');
+      }
     });
-  });
+  }
+
+  return domContentLoadedWrapper(callback);
 }